made arrows shoot in right place and be able to be shot by just holding mouse down

+++ inc/component/functional/RotationComp.hpp
@@ -6,7 +6,7 @@
  */
 struct RotationComp
 {
-	float m_angle;
+	double m_angle;
 };
 
 #endif
